Make work queue push based
The work item queue was pull based, meaning that if the backed had no
work to do, the processor would both spin wheels doing no work as well
as processing the next work item slower as there was a backoff delay to
when there was no work.

Updates the queue to be pull based, calling a function which hangs until
there is either work to do or the context is cancelled.

Signed-off-by: joshvanl <me@joshvanl.dev>